+++ CHANGELOG.md
@@ -14,6 +14,7 @@ AdminNeo 5.6.0
 ### Bugfixes
 
 - SQLite: Preserve STRICT and WITHOUT ROWID in alter table (by @vrana)
+- SQLite: Do not treat INTEGER PRIMARY KEY without AUTOINCREMENT as auto increment (by @vrana)
 
 AdminNeo 5.5.1 (2026-07-20)
 ---------------------------